+++ src/model/Model_Checking.h
@@ -77,8 +77,8 @@ public:
         Full_Data();
         explicit Full_Data(const Data& r);
         Full_Data(const Data& r
-            , const std::map<int /*trans id*/, Model_Splittransaction::Data_Set /*split trans*/ > & splits
-            , const std::map<int /*trans id*/, Model_Taglink::Data_Set /*split trans*/ >& tags);
+            , const std::map<int64 /*trans id*/, Model_Splittransaction::Data_Set /*split trans*/ > & splits
+            , const std::map<int64 /*trans id*/, Model_Taglink::Data_Set /*split trans*/ >& tags);
 
         ~Full_Data();
         wxString ACCOUNTNAME, TOACCOUNTNAME;
@@ -91,9 +91,9 @@ public:
         wxArrayString ATTACHMENT_DESCRIPTION;
         Model_Splittransaction::Data_Set m_splits;
         Model_Taglink::Data_Set m_tags;
-        wxString real_payee_name(int account_id) const;
-        const wxString get_currency_code(int account_id) const;
-        const wxString get_account_name(int account_id) const;
+        wxString real_payee_name(int64 account_id) const;
+        const wxString get_currency_code(int64 account_id) const;
+        const wxString get_account_name(int64 account_id) const;
         bool has_split() const;
         bool has_tags() const;
         bool has_attachment() const;
@@ -193,11 +193,11 @@ public:
     static Model_Checking& instance();
 
 public:
-    bool remove(int id);
-    int save(Data* r);
-    int save(std::vector<Data>& rows);
-    int save(std::vector<Data*>& rows);
-    void updateTimestamp(int id);
+    bool remove(int64 id);
+    int64 save(Data* r);
+    int64 save(std::vector<Data>& rows);
+    int64 save(std::vector<Data*>& rows);
+    void updateTimestamp(int64 id);
 public:
     static const Model_Splittransaction::Data_Set splittransaction(const Data* r);
     static const Model_Splittransaction::Data_Set splittransaction(const Data& r);
@@ -220,23 +220,23 @@ public:
     static STATUS_ID status_id(const wxString& r);
     static STATUS_ID status_id(const Data* r);
     static STATUS_ID status_id(const Data& r);
-    static double amount(const Data* r, int account_id = -1);
-    static double amount(const Data&r, int account_id = -1);
-    static double balance(const Data* r, int account_id = -1);
-    static double balance(const Data& r, int account_id = -1);
-    static double withdrawal(const Data* r, int account_id = -1);
-    static double withdrawal(const Data& r, int account_id);
-    static double deposit(const Data* r, int account_id);
-    static double deposit(const Data& r, int account_id);
-    static double reconciled(const Data* r, int account_id);
-    static double reconciled(const Data& r, int account_id);
+    static double amount(const Data* r, int64 account_id = -1);
+    static double amount(const Data&r, int64 account_id = -1);
+    static double balance(const Data* r, int64 account_id = -1);
+    static double balance(const Data& r, int64 account_id = -1);
+    static double withdrawal(const Data* r, int64 account_id = -1);
+    static double withdrawal(const Data& r, int64 account_id);
+    static double deposit(const Data* r, int64 account_id);
+    static double deposit(const Data& r, int64 account_id);
+    static double reconciled(const Data* r, int64 account_id);
+    static double reconciled(const Data& r, int64 account_id);
     static bool is_locked(const Data* r);
     static bool is_transfer(const wxString& r);
     static bool is_transfer(const Data* r);
     static bool is_deposit(const wxString& r);
     static bool is_deposit(const Data* r);
-    static void getFrequentUsedNotes(std::vector<wxString> &frequentNotes, int accountID = -1);
-    static void getEmptyTransaction(Data &data, int accountID);
+    static void getFrequentUsedNotes(std::vector<wxString> &frequentNotes, int64 accountID = -1);
+    static void getEmptyTransaction(Data &data, int64 accountID);
     static bool getTransactionData(Data &data, const Data* r);
     static void putDataToTransaction(Data *r, const Data &data);
     static bool foreignTransaction(const Data& data);
